Transfer all regInfo to Dto folder and fix dependent classes
