+++ HW6/Valeriy_Satkov/Trading/Traiding.Core.Tests/Traiding.Core.Tests.csproj
@@ -66,7 +66,7 @@
     <Compile Include="ClientsServiceTests.cs" />
     <Compile Include="SharesNumbersServiceTests.cs" />
     <Compile Include="OperationsServiceTests.cs" />
-    <Compile Include="SaleHandlerServiceTests.cs" />
+    <Compile Include="SaleHandlerTests.cs" />
     <Compile Include="SharesServiceTests.cs" />
     <Compile Include="ShareTypesServiceTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
