+++ HW6/Valeriy_Satkov/Trading/Traiding.Core/Dto/BlockedSharesNumberRegistrationInfo.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Traiding.Core.Models;
 
-namespace Traiding.Core.Models
+namespace Traiding.Core.Dto
 {
     public class BlockedSharesNumberRegistrationInfo
     {
