+++ HW6/Valeriy_Satkov/Trading/Traiding.Core/Dto/ClientRegistrationInfo.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Traiding.Core.Models
+namespace Traiding.Core.Dto
 {
     public class ClientRegistrationInfo
     {
