+++ HW6/Valeriy_Satkov/Trading/Traiding.Core/Dto/SharesNumberRegistrationInfo.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Traiding.Core.Models;
+
+namespace Traiding.Core.Dto
+{
+    public class SharesNumberRegistrationInfo
+    {
+        public ClientEntity Client { get; set; }
+        public ShareEntity Share { get; set; }
+        public int Number { get; set; }
+    }
+}
