--- HW6/Valeriy_Satkov/Trading/Traiding.Core/Models/ShareTypeRegistrationInfo.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Traiding.Core.Models
-{
-    public class ShareTypeRegistrationInfo
-    {
-        public string Name { get; set; }
-        public decimal Cost { get; set; }
-        public bool Status { get; set; }
-    }
-}
