+++ HW6/Valeriy_Satkov/Trading/Traiding.Core/Services/BlockedSharesNumbersService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Traiding.Core.Dto;
 using Traiding.Core.Repositories;
 using Traiding.Core.Models;
