+++ HW6/Valeriy_Satkov/Trading/Traiding.Core/Traiding.Core.csproj
@@ -48,18 +48,18 @@
     <Compile Include="Models\BalanceEntity.cs" />
     <Compile Include="Models\BlockedMoneyEntity.cs" />
     <Compile Include="Models\BlockedSharesNumberEntity.cs" />
-    <Compile Include="Models\BalanceRegistrationInfo.cs" />
-    <Compile Include="Models\BlockedMoneyRegistrationInfo.cs" />
-    <Compile Include="Models\BlockedSharesNumberRegistrationInfo.cs" />
-    <Compile Include="Models\SharesNumberRegistrationInfo.cs" />
-    <Compile Include="Models\ShareRegistrationInfo.cs" />
+    <Compile Include="Dto\BalanceRegistrationInfo.cs" />
+    <Compile Include="Dto\BlockedMoneyRegistrationInfo.cs" />
+    <Compile Include="Dto\BlockedSharesNumberRegistrationInfo.cs" />
+    <Compile Include="Dto\SharesNumberRegistrationInfo.cs" />
+    <Compile Include="Dto\ShareRegistrationInfo.cs" />
     <Compile Include="Models\OperationEntity.cs" />
     <Compile Include="Models\SharesNumberEntity.cs" />
     <Compile Include="Models\ShareEntity.cs" />
     <Compile Include="Models\ShareTypeEntity.cs" />
-    <Compile Include="Models\ShareTypeRegistrationInfo.cs" />
+    <Compile Include="Dto\ShareTypeRegistrationInfo.cs" />
     <Compile Include="Models\ClientEntity.cs" />
-    <Compile Include="Models\ClientRegistrationInfo.cs" />
+    <Compile Include="Dto\ClientRegistrationInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Repositories\IOperationTableRepository.cs" />
     <Compile Include="Repositories\IBlockedSharesNumberTableRepository.cs" />
